test(old_arch_compatibility): Skip test if NEXT_GEN is set
- The `NEXT_GEN` environment variable is used to indicate that the test should be skipped.
- This allows for selective execution of tests when working on new features or architectures.

+++ tests/integration_tests/old_arch_compatibility/run.sh
@@ -12,6 +12,10 @@ CDC_BINARY=cdc.test
 SINK_TYPE=$1
 
 function run() {
+	if [ "$NEXT_GEN" = 1 ]; then
+		exit 0
+	fi
+
 	rm -rf $WORK_DIR && mkdir -p $WORK_DIR
 
 	start_tidb_cluster --workdir $WORK_DIR
